moved dispositivo interface to a separete file in domain
