+++ src/domain/dispositivo/dispositivo-interface.ts
@@ -1,0 +1,10 @@
+interface Dispositivo {
+  id: number
+  nome: string
+  descricao: string
+  endereco: string
+  local: string
+  gatewayId: number | null
+}
+
+export default Dispositivo
